move header from appshell to each page
