+++ src/components/CreateFamily/FamilyButton/CreateFamilyButton.tsx
@@ -1,7 +1,6 @@
 import { PlusIcon } from '@camp/design';
 import { messages } from '@camp/messages';
 import { Button } from '@mantine/core';
-import { Link } from '@tanstack/react-location';
 
 interface Props {
   onClick?: React.MouseEventHandler<HTMLButtonElement>;
@@ -9,22 +8,20 @@ interface Props {
 
 export const CreateFamilyButton = ({ onClick }: Props) => {
   return (
-    <Link to="/create-draft-family">
-      <Button
-        variant="outline"
-        size="sm"
-        id="create-draft-family"
-        leftIcon={<PlusIcon width="16" height="16" />}
-        onClick={onClick}
-        sx={theme => ({
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          '&:hover': {
-            backgroundColor: theme.colors.primarySubtle[6],
-          },
-        })}
-      >
-        {messages.families.create}
-      </Button>
-    </Link>
+    <Button
+      variant="outline"
+      size="sm"
+      id="create-draft-family"
+      leftIcon={<PlusIcon width="16" height="16" />}
+      onClick={onClick}
+      sx={theme => ({
+        // eslint-disable-next-line @typescript-eslint/naming-convention
+        '&:hover': {
+          backgroundColor: theme.colors.primarySubtle[6],
+        },
+      })}
+    >
+      {messages.families.create}
+    </Button>
   );
 };
