+++ src/components/appShell/index.ts
@@ -1,1 +1,2 @@
 export * from './AppShell';
+export * from './Breadcrumbs';
